Move code patch models into dedicated module
Move LogicalQubit and CodePatch from the lattice surgery primitives
module into codepatch.py, along with their dependencies and tests.

Update architecture imports and package exports while preserving the
existing public API and resource-count regression coverage.
